myApp/server/requirements.txt
appdirs==1.4.3
click==6.7
Django==1.10.6
django-filter==1.0.1
djangorestframework==3.6.2
djangorestframework-jwt==1.9.0
first==2.0.1
Markdown==2.6.8
packaging==16.8
pip-tools==1.8.0
PyJWT==1.4.2
pyparsing==2.2.0
six==1.10.0
